first iteration of integration tests
