+++ e2e/integration/spec/authoring-api/test-data/bulk1.zip
