+++ e2e/integration/spec/authoring-api/test-data/invalidRules.zip
