+++ e2e/integration/spec/authoring-api/test-data/notZip.zip
